requirements.txt
certifi==2019.9.11
cffi==1.13.2
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.3
olefile==0.46
Pillow==6.2.1
pycparser==2.19
six==1.13.0
torch==1.3.1
torchvision==0.4.2
